Rename: rename rythm_game to rhythm
