+++ rhythm.h
